# docs/openapi.yaml
openapi: 3.0.1
info:
  title: 콘서트 예약 서비스 API
  description: "대기열 관리, 좌석 점유, 결제 기능을 제공하는 콘서트 예약 서비스"
  contact:
    name: 박서희
    email: sh940311@gmail.com
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local 개발 서버
tags:
  - name: Queue
    description: 대기열 토큰 조회/발급
  - name: Seats
    description: 좌석 조회/점유
  - name: Balance
    description: 잔액 조회/충전
  - name: Reservations
    description: 결제 및 예약
paths:
  /api/v1/reservations/concerts/{concertId}/seats/hold:
    post:
      tags:
        - Reservations
      summary: 좌석 점유 요청
      operationId: holdSeats
      parameters:
        - name: Queue-Token
          in: header
          description: 대기열 토큰
          required: true
          schema:
            type: string
            format: encoded
            example: abcac10b-58cc-4372-a567-0e02b2c3d479
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HoldSeatRequest"
        required: true
      responses:
        "200":
          description: 점유 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                code: SUCCESS
                message: 좌석 점유 성공
                data:
                  seatId: 1
                  seatNumber: 1
                  expiresAt: 2025-07-20T19:12:34Z
  /api/v1/reservations/:
    post:
      tags:
        - Reservations
      summary: 결제 및 예약 확정
      operationId: confirmedReservation
      parameters:
        - name: Queue-Token
          in: header
          description: 대기열 토큰
          required: true
          schema:
            type: string
            format: encoded
            example: abcac10b-58cc-4372-a567-0e02b2c3d479
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmReservationRequest"
        required: true
      responses:
        "200":
          description: 결제 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                code: SUCCESS
                message: 예약이 성공적으로 완료되었습니다.
                data:
                  concertId: 1
                  seatId: 1
                  price: 130000
  /api/v1/queue/token/:
    get:
      tags:
        - Queue
      summary: 대기 번호 및 토큰 상태 조회
      operationId: getTokenStatus
      parameters:
        - name: Queue-Token
          in: header
          description: 대기열 토큰
          required: true
          schema:
            type: string
            format: encoded
            example: abcac10b-58cc-4372-a567-0e02b2c3d479
      responses:
        "200":
          description: 토큰 ACTIVE 혹은 WAITING 상태
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                code: SUCCESS
                message: 토큰 상태 조회 성공
                data:
                  status: ACTIVE
                  position: 0
                  expiresAt: 2025-07-20T19:12:34Z
    post:
      tags:
        - Queue
      summary: 대기열 토큰 발급
      operationId: createToken
      parameters:
        - name: Client-Id
          in: header
          description: 클라이언트 식별자
          required: true
          schema:
            type: number
      responses:
        "201":
          description: 토큰 발급 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                code: SUCCESS
                message: 토큰이 성공적으로 발급되었습니다.
                data:
                  token: abcac10b-58cc-4372-a567-0e02b2c3d479
                  status: ACTIVE
                  position: 0
                  expiresAt: 2025-07-20T19:12:34Z
  /api/v1/balance/charge:
    post:
      tags:
        - Balance
      summary: 잔액 충전
      operationId: chargeBalance
      parameters:
        - name: Client-Id
          in: header
          description: 클라이언트 식별자
          required: true
          schema:
            type: number
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChargeBalanceRequest"
        required: true
      responses:
        "200":
          description: 충전 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                code: SUCCESS
                message: 충전 성공
                data:
                  balance: 100000
  /api/v1/reservations/concerts:
    get:
      tags:
        - Reservations
      summary: 예약 가능한 콘서트 목록 조회
      operationId: getConcerts
      parameters:
        - name: Queue-Token
          in: header
          description: 대기열 토큰
          required: true
          schema:
            type: string
            format: encoded
            example: abcac10b-58cc-4372-a567-0e02b2c3d479
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                code: SUCCESS
                message: 요청이 성공했습니다.
                data:
                  - concertId: 1
                    concertDateTime: 2023-10-01T19:00:00
                    concertVenue: 서울 올림픽공원 체조경기장
                    concertTitle: 2023 HH+ Concert
                    isAvailable: true
                  - concertId: 2
                    concertDateTime: 2023-10-02T19:00:00
                    concertVenue: 서울 올림픽공원 주경기장
                    concertTitle: 2023 Jazz Concert
                    isAvailable: false
  /api/v1/reservations/concerts/{concertId}/seats:
    get:
      tags:
        - Reservations
      summary: 특정 콘서트의 좌석 목록 조회
      operationId: getSeats
      parameters:
        - name: Queue-Token
          in: header
          description: 대기열 토큰
          required: true
          schema:
            type: string
            format: encoded
            example: abcac10b-58cc-4372-a567-0e02b2c3d479
        - name: concertId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                code: SUCCESS
                message: 요청이 성공했습니다.
                data:
                  concertId: 1
                  availableSeats:
                    - seatId: 1
                      seatNumber: 1
                      price: 130000
                      isAvailable: true
                    - seatId: 2
                      seatNumber: 2
                      price: 150000
                      isAvailable: false
  /api/v1/balance/:
    get:
      tags:
        - Balance
      summary: 잔액 조회
      operationId: getBalance
      parameters:
        - name: Client-Id
          in: header
          description: 클라이언트 식별자
          required: true
          schema:
            type: number
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
              example:
                code: SUCCESS
                message: 잔액 조회 성공
                data:
                  balance: 50000
components:
  schemas:
    HoldSeatRequest:
      type: object
      properties:
        seatHoldUuid:
          type: string
        concertId:
          type: integer
          format: int64
        seatId:
          type: integer
          format: int64
      description: 좌석 점유 요청
    ApiResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: object
      description: API 응답 형식
    ConfirmReservationRequest:
      type: object
      properties:
        reservationUuid:
          type: string
        seatId:
          type: integer
          format: int64
      description: 예약 확정 요청
    ChargeBalanceRequest:
      type: object
      properties:
        amount:
          type: integer
          format: int64
      description: 잔액 충전 요청